added optional action & action views templates
Added the possibility to have custom action templates and view templates
for controllers.

Generate the controller as normal (i.e. bob c test add edit delete)
when generating the controller and it's views, bob will look for a
/bob/controller/action.#action_name#.tpl file for specific actions and
view.#action_name#.tpl for the views.

This will allow you for instance to always fire a custom event in your
controllers when the delete action is called